Create test cases, step 1, no edges created in edges-test.json
